requirements.txt
# Python 3.10+ required
requests
beautifulsoup4
python-dotenv
